[msbuild] ongoing ITaskItem input parameter handling improvements.
Right now all the input parameters are computed based on string, and any
metadata items are simply chopped out when it is passed to a Task.
To fix this, we need to replace ExpandString() use with more complicated
evaluator that takes ITaskItem into consideration.

(It is not complete as of now; evaluation method needs to be replaced and
any regressions need to be fixed.)