[Managed.Windows.Forms] Fix 'minimum' default value of NumericUpDown
The 'minimum' default value of NumericUpDown was set to 0.0M, which caused the ToString() to format it as "0.0" instead of "0" as it does on .NET.
A similar change was done for the 'maximum' default value already in f6c8d48a3aed377b8a11e38c96c0d4c774664279.

Commit licensed under MIT/X11.