Print pid_t's portably in mono/metadata/attach.c
The pid_t type is required by POSIX to be a signed integer type (and that it
isn't wider than long in at least one programming environment). Mono has a
bunch of issues with 64-bit pids at the present, this is a start. Cast the
pid to the largest available signed type before printing it.

This change is released under the MIT/X11 license.