Use mono_strtod () instead of strtod () on ARM.
The former used to miscompile back in 2005, but this no
longer appears to be the case. Further, using strtod ()
is problematic on Linux/ARM systems with many non-US
locales that use a comma as the decimal separator.

In particular, this breaks almost all compilation with
MCS because it can't parse floating point literals.

Based on investigation by Damien Diederen <dd@crosstwine.com>.