[sgen] File size limit for binary protocols.
We can now pass a file size limit to the `binary-protocol` option.  It will
create numbered binary protocol files, and only keep the last two around.
This is useful for cases where the only way to reproduce a bug is through
a long run that would otherwise create a protocol that's much too big.